requirements.txt
aiofiles
aiohttp
aiohttp_session
cryptography
pyrofork
python-dotenv
TgCrypto-pyrofork

motor
pymongo
py-tgcalls
uvloop

# Save-Restricted-Content-Bot features
pytz
yt-dlp
Pillow